requirements.txt
pip==20.2.3
matplotlib==3.3.2
speedtest-cli==2.1.2
requests==2.24.0